// src/app/models/product.model.ts
export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: {
    rate: number;
    count: number;
  };
  discount?: number; // Optional discount percentage
  quantity?: number; // Optional quantity for cart items
}